pyproject.toml
[project]
name = "qss4"
version = "0.1.0"
description = "end to end PQ safe file storage solution"
requires-python = ">=3.11"
dependencies = [
    "click>=8.2.1",
    "email-validator>=2.2.0",
    "eth-account>=0.13.7",
    "flask-jwt-extended>=4.7.1",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psutil>=7.0.0",
    "psycopg2-binary>=2.9.10",
    "redis>=6.4.0",
    "sqlalchemy>=2.0.42",
    "web3>=7.13.0",
    "werkzeug>=3.1.3",
    "zstandard>=0.23.0",
    "flask-login>=0.6.3",
    "cryptography>=45.0.6",
    "pydantic>=2.11.7",
    "keyring>=25.6.0",
    "python-magic>=0.4.27",
    "requests>=2.32.4",
    "rich>=14.1.0",
    "pydantic-settings>=2.10.1",
    "kyber-py>=1.0.1",
]
